jack1: new package
This patch adds a new jack1 package alongside the existing jack2.

While jack1 and jack2 are two equivalent implementations of the same
protocol, they differ in a few details and both of them are being
actively used (jack2 is not considered to be a replacement of jack1).

For more information, see:
https://github.com/jackaudio/jackaudio.github.com/wiki/Q_difference_jack1_jack2

Signed-off-by: Adam Heinrich <adam@adamh.cz>